requirements.txt
# ========================
# Production Dependencies
# ========================

# Веб-фреймворк
Django==5.2.7
django-filter==25.2

# Веб-сервер
# gunicorn

# Драйвер PostgreSQL
psycopg[binary]==3.2.10

# Переменные окружения
python-decouple==3.8

# Обработка изображений (для ImageField)
pillow==11.3.0

# Валидация и анализ телефонных номеров
phonenumberslite==9.0.16

# Логирование
# loguru

# Мониторинг
# sentry-sdk
